requirements.txt
torch
transformers
streamlit
chromadb
faiss-cpu
langchain
langchain-community
nltk
spacy
pdfplumber
openpyxl
python-docx
pandas
sentence-transformers
huggingface_hub